LayoutTests/platform/qt/fast/events/continuous-platform-wheelevent-in-scrolling-div-expected.txt

FAIL event.wheelDeltaY should be -24000. Was 0.
PASS event.wheelDeltaX is window.expectedScrollLeft*-120
PASS event.wheelDelta is window.expectedScrollLeft*-120
PASS event.wheelDeltaY is window.expectedScrollTop*-120
FAIL event.wheelDeltaX should be -12000. Was 0.
PASS event.wheelDelta is window.expectedScrollTop*-120
PASS div.scrollTop is window.expectedScrollTop
PASS div.scrollLeft is window.expectedScrollLeft
